Rename `exoshell::eventstream` → `exoshell::eventq`.
