+++ exoshell/src/mainloop.rs
@@ -1,10 +1,10 @@
-use crate::{eventstream, UI};
+use crate::{eventq, UI};
 
 pub(crate) async fn main_loop() -> anyhow::Result<()> {
     use crate::cleanup::CleanupWith;
     use crate::Runner;
 
-    let (r, s) = eventstream::init();
+    let (r, s) = eventq::init();
 
     let mut ui = UI::new(Runner::from(s.clone()))?;
     main_loop_inner(r, &mut ui)
@@ -14,10 +14,7 @@ pub(crate) async fn main_loop() -> anyhow::Result<()> {
     Ok(())
 }
 
-pub(crate) async fn main_loop_inner(
-    mut evr: eventstream::Reader,
-    ui: &mut UI,
-) -> anyhow::Result<()> {
+pub(crate) async fn main_loop_inner(mut evr: eventq::Reader, ui: &mut UI) -> anyhow::Result<()> {
     use crate::AppEvent::*;
 
     while let Some(event) = evr.next().await {
